flowsint-api/requirements.txt
fastapi
redis==5.0.1
supabase==2.13.0
celery==5.3.6
redis==5.0.1
holehe
dotenv
uvicorn
sherlock_project
python-jose[cryptography]
python-dotenv
requests
git+https://github.com/soxoj/maigret
python-whois
pydig
ghunt
ignorant
hibpwned
phonenumbers